Use mix of GraphQL and processing progress API to set document title
